Call Error() when aggregating error structs with no public fields.
Previously, aggregatestruct was relying on reflect.Value.NumField() to
determine whether to try calling an Error() method on a type.  This
counted unexported fields, like the members of the stdlib errorString
type, causing them to not aggregate anything.

Instead, count the number of public fields aggregated.  If that count is
zero, then try to see if it is an error.