Remove Google App Engine specific support.
With the release of Go 1.11 support on Google App Engine, the
"appengine" tag will no longer be applied to builds.  Code is expected
to not have special cases for App Engine.

For example, instead of using the appengine/log package, code is
expected to just use the plain standard library "log" package.

See the migration guide for more information:

  https://cloud.google.com/appengine/docs/standard/go111/go-differences